fix: moved plugins
