+++ src/mindmap-view.ts
@@ -16,7 +16,7 @@ import { MD_VIEW_TYPE, MM_VIEW_TYPE } from "./constants";
 import ObsidianMarkmap from "./obsidian-markmap-plugin";
 import { createSVG, getComputedCss, removeExistingSVG } from "./markmap-svg";
 import { copyImageToClipboard } from "./copy-image";
-import { htmlEscapePlugin, checkBoxPlugin } from "./custom-plugins";
+import { htmlEscapePlugin, checkBoxPlugin } from "./plugins";
 import { MindMapSettings } from "./settings";
 import { FrontmatterOptions } from "./@types/models";
