+++ src/plugins/index.ts
@@ -1,0 +1,2 @@
+export { htmlEscapePlugin } from "./html-escape";
+export { checkBoxPlugin } from "./checkbox";
